+++ src/pages/onboarding/components/FinishStep/index.tsx
