+++ src/pages/onboarding/components/GenreChip/index.css.ts
