+++ src/pages/onboarding/components/GenreStep/index.css.ts
