+++ src/pages/onboarding/components/InfoStep/index.css.ts
