+++ src/pages/onboarding/index.css.ts
