+++ src/routes/router.tsx
@@ -9,6 +9,7 @@ import InstructorRegister from '@/pages/instructorRegister';
 import Login from '@/pages/login';
 import MyPage from '@/pages/mypage';
 import MyPageReservation from '@/pages/mypage/mypageReservation';
+import Onboarding from '@/pages/onboarding';
 import Reservation from '@/pages/reservation';
 import Search from '@/pages/search';
 import { ROUTES_CONFIG } from './routesConfig';
@@ -21,6 +22,10 @@ export const router = createBrowserRouter([
   {
     path: ROUTES_CONFIG.login.path,
     element: <Login />,
+  },
+  {
+    path: ROUTES_CONFIG.onboarding.path,
+    element: <Onboarding />,
   },
   {
     path: ROUTES_CONFIG.search.path,
